[AIP-21] Move sensors.s3_key_sensor airflow.sensors.s3_key_sensor
