+++ docs/operators-and-hooks-ref.rst
@@ -406,7 +406,7 @@ These integrations allow you to perform various operations within the Amazon Web
        :mod:`airflow.providers.amazon.aws.operators.s3_copy_object`,
        :mod:`airflow.providers.amazon.aws.operators.s3_delete_objects`,
        :mod:`airflow.providers.amazon.aws.operators.s3_list`
-     - :mod:`airflow.sensors.s3_key_sensor`,
+     - :mod:`airflow.providers.amazon.aws.sensors.s3_key`,
        :mod:`airflow.sensors.s3_prefix_sensor`
 
 Transfer operators and hooks
