+++ scripts/ci/pylint_todo.txt
@@ -175,7 +175,7 @@
 ./airflow/sensors/http_sensor.py
 ./airflow/providers/apache/hive/sensors/metastore_partition.py
 ./airflow/providers/apache/hive/sensors/named_hive_partition.py
-./airflow/sensors/s3_key_sensor.py
+./airflow/providers/amazon/aws/sensors/s3_key.py
 ./airflow/sensors/s3_prefix_sensor.py
 ./airflow/sensors/sql_sensor.py
 ./airflow/sensors/time_delta_sensor.py
